Continued rework / integration for ACE
- fixed vitals HUD still showing when doing medical actions

- removed vitals HUD when accessing target rather than self for medical

- added dispatch action for small game animals, you can now snap their neck with ACE interaction

- fixed near check for animals field dressing / processing, new function now passes animal object directly to skin function
