+++ README.md
@@ -55,7 +55,7 @@ We trust that this mod will serve as a valuable tool in realizing your vision fo
 
 ## Getting started
 
-Misery requires Arma 3 and the latest version of <a href="https://github.com/CBATeam/CBA_A3/releases">CBA A3</a>.
+Misery requires Arma 3 and the latest version of <a href="https://github.com/CBATeam/CBA_A3/releases">CBA A3</a> and <a href="https://github.com/acemod/ACE3/releases">ACE3</a>.
 
 ## Contributing to Misery
