+++ addons/actions/CfgEventHandlers.hpp
@@ -1,5 +1,5 @@
-class Extended_PreInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_SCRIPT(XEH_preInit));
-    };
-};
+// class Extended_PreInit_EventHandlers {
+//     class ADDON {
+//         init = QUOTE(call COMPILE_SCRIPT(XEH_preInit));
+//     };
+// };
