+++ addons/forestry/XEH_postInit.sqf
@@ -4,25 +4,25 @@ if (GVAR(woodCollection)) then {
 
     GVAR(gatheredPositions) = [];
 
-    [
-        "forestryCollectWood_menu",
+    private _forestryCollectWood = [
+        QGVAR(forestryCollectWood_menu),
         localize ECSTRING(common,CollectWood),
-        {([player] call EFUNC(common,nearTree)) select 0},
+        QPATHTOEF(icons,data\trees_ca.paa),
         {
-            [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
             [] call FUNC(forageTreeAction);
         },
-        "",
-        QPATHTOEF(icons,data\trees_ca.paa),
-        ""
-    ] call EFUNC(actions,addAction);
-
-    [
-        "forestrySplitLog_menu",
-        localize ECSTRING(common,SplitWoodLog),
-        {[[QCLASS(woodenlog)]] call EFUNC(common,hasItem)},
         {
-            [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
+            insideBuilding player isEqualTo 0 && ([player] call EFUNC(common,nearTree)) select 0
+        }
+    ] call ace_interact_menu_fnc_createAction;
+
+    [player, 1, ["ACE_SelfActions"], _forestryCollectWood] call ace_interact_menu_fnc_addActionToObject;
+
+    private _forestrySplitLog = [
+        QGVAR(forestrySplitLog_menu),
+        localize ECSTRING(common,SplitWoodLog),
+        QPATHTOEF(icons,data\axe_ca.paa),
+        {
             if !([[QCLASS(woodenlog)]] call EFUNC(common,hasItem)) exitWith {
                 [QEGVAR(common,tileText), localize ECSTRING(common,NoWoodenLogsForSplitting)] call CBA_fnc_localEvent;
             };
@@ -36,10 +36,12 @@ if (GVAR(woodCollection)) then {
                 [QEGVAR(common,tileText), localize ECSTRING(common,NoWoodAxeOrChainsawNoti)] call CBA_fnc_localEvent;
             };
         },
-        "",
-        QPATHTOEF(icons,data\axe_ca.paa),
-        ""
-    ] call EFUNC(actions,addAction);
+        {
+            [[QCLASS(woodenlog)]] call EFUNC(common,hasItem)
+        }
+    ] call ace_interact_menu_fnc_createAction;
+
+    [player, 1, ["ACE_SelfActions"], _forestrySplitLog] call ace_interact_menu_fnc_addActionToObject;
 };
 
 if (GVAR(foraging)) then {
@@ -47,42 +49,39 @@ if (GVAR(foraging)) then {
     GVAR(tinderPositions) = [];
     GVAR(digPositions) = [];
 
-    [
-        "foraging_menu",
+    private _foragingMenu = [
+        QGVAR(foraging_menu),
         localize ECSTRING(common,Forage),
-        {call FUNC(canForage)},
-        {
-            player setVariable [QEGVAR(actions,currentParentID), "foraging_menu"];
-            call EFUNC(actions,displayActions);
-        },
-        "",
         QPATHTOEF(icons,data\leaf_ca.paa),
-        ""
-    ] call EFUNC(actions,addAction);
+        {},
+        {
+            call FUNC(canForage)
+        }
+    ] call ace_interact_menu_fnc_createAction;
 
-    [
-        "forageWorms_menu",
+    [player, 1, ["ACE_SelfActions"], _foragingMenu] call ace_interact_menu_fnc_addActionToObject;
+
+    private _foragingWormsMenu = [
+        QGVAR(foragingWorms_menu),
         localize ECSTRING(common,ForageDigForWorms),
-        {player getVariable [QEGVAR(actions,currentParentID), ""] isEqualTo "foraging_menu"},
-        {
-            [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
-            [] call FUNC(digForWorms);
-        },
-        "foraging_menu",
         QPATHTOEF(icons,data\shovel_ca.paa),
-        ""
-    ] call EFUNC(actions,addAction);
+        {
+            [] call FUNC(digForWorms)
+        },
+        {true}
+    ] call ace_interact_menu_fnc_createAction;
 
-    [
-        "forageTinder_menu",
+    [player, 1, ["ACE_SelfActions", QGVAR(foraging_menu)], _foragingWormsMenu] call ace_interact_menu_fnc_addActionToObject;
+
+    private _foragingTinderMenu = [
+        QGVAR(foragingTinder_menu),
         localize ECSTRING(common,ForageSearchForTinder),
-        {player getVariable [QEGVAR(actions,currentParentID), ""] isEqualTo "foraging_menu"},
+        QPATHTOEF(icons,data\leaf_ca.paa),
         {
-            [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
-            [] call FUNC(searchForTinder);
+            [] call FUNC(searchForTinder)
         },
-        "foraging_menu",
-        QPATHTOEF(icons,data\shovel_ca.paa),
-        ""
-    ] call EFUNC(actions,addAction);
+        {true}
+    ] call ace_interact_menu_fnc_createAction;
+
+    [player, 1, ["ACE_SelfActions", QGVAR(foraging_menu)], _foragingTinderMenu] call ace_interact_menu_fnc_addActionToObject;
 };
