+++ addons/furniture/CfgEventHandlers.hpp
@@ -1,11 +1,11 @@
-class Extended_PreInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_SCRIPT(XEH_preInit));
-    };
-};
+// class Extended_PreInit_EventHandlers {
+//     class ADDON {
+//         init = QUOTE(call COMPILE_SCRIPT(XEH_preInit));
+//     };
+// };
 
-class Extended_PostInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_SCRIPT(XEH_postInit));
-    };
-};
+// class Extended_PostInit_EventHandlers {
+//     class ADDON {
+//         init = QUOTE(call COMPILE_SCRIPT(XEH_postInit));
+//     };
+// };
