+++ addons/geiger/XEH_postInit.sqf
@@ -6,19 +6,18 @@ if (hasInterface) then {
 
 if !(isServer) exitWith {};
 
-[
-    "geiger_menu",
+private _addBatterytoGeiger = [
+    QGVAR(geiger_menu),
     "Add lithium battery to geiger",
+    QPATHTOEF(icons,data\battery_charging_ca.paa),
+    {
+        call FUNC(batteries)
+    },
     {
         [[QCLASS(lithiumBattery), QCLASS(geiger_NoBattery)]] call EFUNC(common,hasItem)
-    },
-    {
-        [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
-        call FUNC(batteries);
-    },
-    "",
-    QPATHTOEF(icons,data\battery_charging_ca.paa),
-    ""
-] call EFUNC(actions,addAction);
+    }
+] call ace_interact_menu_fnc_createAction;
+
+[player, 1, ["ACE_SelfActions"], _addBatterytoGeiger] call ace_interact_menu_fnc_addActionToObject;
