+++ addons/radiation/XEH_postInit.sqf
@@ -8,18 +8,19 @@
     if (hasInterface) then {
         [QGVAR(radiationEvent), FUNC(process)] call CBA_fnc_addEventHandler;
 
-        [
-            "leadContainers_menu",
+        private _leadContainer = [
+            QGVAR(leadContainers_menu),
             localize ECSTRING(common,StoreArtifact),
-            {[[QCLASS(leadContainer_Open)]] call EFUNC(common,hasItem) && [[MACRO_ARTIFACTS]] call EFUNC(common,hasItem)},
-            {
-            [QEGVAR(common,exitGui)] call CBA_fnc_localEvent;
-            call FUNC(storeArtifact);
-            },
-            "",
             QPATHTOEF(icons,data\package_open_ca.paa),
-            ""
-        ] call EFUNC(actions,addAction);
+            {
+                call FUNC(storeArtifact)
+            },
+            {
+                [[QCLASS(leadContainer_Open)]] call EFUNC(common,hasItem) && [[MACRO_ARTIFACTS]] call EFUNC(common,hasItem)
+            }
+        ] call ace_interact_menu_fnc_createAction;
+
+        [player, 1, ["ACE_SelfActions"], _leadContainer] call ace_interact_menu_fnc_addActionToObject;
 
 
         // Reactivate Geiger if picking up active one:
