+++ addons/vitals/CfgEventHandlers.hpp
@@ -4,5 +4,9 @@ class Extended_PreInit_EventHandlers {
     };
 };
 
-
+class Extended_PostInit_EventHandlers {
+    class ADDON {
+        init = QUOTE(call COMPILE_SCRIPT(XEH_postInit));
+    };
+};
